[8009] move tito events into own app
