+++ network-api/networkapi/events/urls.py
@@ -1,6 +1,6 @@
 from django.urls import re_path
 
-from networkapi.mozfest.views import tito_ticket_completed
+from .views import tito_ticket_completed
 
 urlpatterns = [
     re_path('^ticket-completed/?$', tito_ticket_completed, name='tito-ticket-completed'),
