+++ network-api/networkapi/events/views.py
