+++ network-api/networkapi/urls.py
@@ -86,7 +86,7 @@ urlpatterns = list(filter(None, [
     re_path(r'^api/people/', include('networkapi.people.urls')),
     re_path(r'^environment.json', EnvVariablesView.as_view()),
     re_path(r'^help/', review_app_help_view, name='Review app help'),
-    re_path(r'^tito/', include('networkapi.mozfest.urls')),
+    re_path(r'^tito/', include('networkapi.events.urls')),
 
     # Wagtail CMS routes
     re_path(
